Fixed age. Files working

+++ src/Users/Streamer.cpp
@@ -10,8 +10,8 @@
 
 Streamer::Streamer(std::string name, std::string nickName, const Date &birthDate) :
         User(name, std::move(nickName), birthDate) {
-    if(age <= minimumAge)
-        throw RestrictedAgeException(name, age, minimumAge);
+    if(age() <= minimumAge)
+        throw RestrictedAgeException(name, (int) age(), minimumAge);
 }
